03_System/pyproject.toml
[project]
name = "theology-ai-lab"
version = "2.0.0"
description = "Theological research vector database with MCP server integration"
requires-python = ">=3.11"
readme = "../README.md"

dependencies = [
    # Core - Vector Database
    "chromadb>=1.3.0",

    # Core - Embeddings
    "sentence-transformers>=5.0.0",
    "transformers>=4.40.0",

    # Core - Deep Learning (CPU only for portability)
    "torch>=2.2.0",

    # Core - Essential
    "numpy>=1.26.0,<2.0",

    # MCP Integration
    "mcp>=1.0.0",

    # PDF Processing
    "pypdf>=5.0.0",
    "langchain-text-splitters>=0.3.0",
    "tiktoken>=0.8.0",

    # OCR Support
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

agents = [
    "google-genai>=0.3.0",
    "playwright>=1.49.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
